// src/tokens/typography/types.ts
type TypographyTokenValue = Record<string, string>

export interface TypographyTokenGroup {
  base: TypographyTokenValue
  h1: TypographyTokenValue
  h2: TypographyTokenValue
  h3: TypographyTokenValue
  h4: TypographyTokenValue
  h5: TypographyTokenValue
  h6: TypographyTokenValue
  p1: TypographyTokenValue
  p1compact: TypographyTokenValue
  p2: TypographyTokenValue
  p3: TypographyTokenValue
}